move get_definition_from in
